chess-service/src/chess_service/main.py
import uvicorn
from dotenv import load_dotenv
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware

from chess_service.routes.routes import router

load_dotenv()

app = FastAPI(title="chess-service")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

app.include_router(router)


if __name__ == "__main__":
    uvicorn.run("chess.main:app", host="0.0.0.0", port=8000, reload=True)
